Split NotebookClient to its own file
